TaskController: Starts removing all the logic from the controller file and moving it to a Services folder, to achieve that implements the use of dependency injection to import the service class on the controller.
TaskService: Changes the index name of the user index in the request to email to ensure data consistency and code maintainability, also creates a condition to check if the user that's being used to create the task exists on the data base, if not performs a redirect with an error, plus takes advantage of the relationships of the models to assign new values to the task model related to the user_id, and not to mention that I've created an interface to keep consistency on future classes.

StoreTaskRequest: Designate the responsibility of validating the request structure to an independent file and following in that way the Laravel Conventions.

Models/Task.php: Replaces the hardcoded relative path with PHP class constant to dynamically reference the User model.

TaskMigration: Improve and reduce the verbosity by specifying in 1 line the relationship and the constraint that relations the tasks table with the users table